+++ src/app/tracker/page.tsx
@@ -36,8 +36,6 @@ const itemList = remnantItems
     discovered: false,
   }))
 
-const totalItems = itemList.length
-
 export default function Page() {
   const isClient = useIsClient()
 
@@ -50,6 +48,7 @@ export default function Page() {
   const { discoveredItemIds } = itemTrackerStorage
 
   const { filteredItems, handleUpdateFilters } = useFilteredItems(itemList)
+  const totalItems = filteredItems.length
 
   // get response after save file upload
   const [uploadFormResponse, formAction] = useFormState(parseSaveFile, {
@@ -146,10 +145,15 @@ export default function Page() {
   }, [setDiscoveredItemIds, filteredItems])
 
   // Provider the tracker progress
-  const discoveredCount = discoveredItemIds.length
+  const discoveredCount = filteredItems.reduce((acc, item) => {
+    if (discoveredItemIds.includes(item.id)) return acc + 1
+    return acc
+  }, 0)
   const discoveredPercent = Math.round((discoveredCount / totalItems) * 100)
   const progress = isClient
-    ? `${discoveredCount} / ${totalItems} (${discoveredPercent}%)`
+    ? `${discoveredCount} / ${totalItems} (${
+        isNaN(discoveredPercent) ? '0' : discoveredPercent
+      }%)`
     : 'Calculating...'
 
   const handleShowItemInfo = (itemId: string) => {